picoxcell: move some startup functions from board to picoxcell
Some common startup functions have been moved from board specific
code to picocxcell specific code.
