+++ arch/arm/cpu/arm926ejs/picoxcell/Makefile
@@ -44,6 +44,7 @@ LIB	= $(obj)lib$(SOC).o
 
 SOBJS-y := reset.o
 
+COBJS-y += cpu.o
 COBJS-y += emac.o
 COBJS-y += timer.o
 COBJS-y += spi.o
