+++ board/picochip/pc7302/pc7302.c
@@ -17,11 +17,9 @@
 
 /* Includes ---------------------------------------------------------------- */
 #include <common.h>
-#include <asm/arch/timer.h>
-#include <asm/arch/utilities.h>
 #include <asm/arch/mux.h>
+#include <asm/arch/utilities.h>
 #include <asm/arch/picoxcell_gpio.h>
-#include <netdev.h>
 
 /* Macros ------------------------------------------------------------------ */
 
@@ -29,14 +27,6 @@
 DECLARE_GLOBAL_DATA_PTR;
 
 /* Prototypes--------------------------------------------------------------- */
-/*!
- *
- * Start a timer in free running mode
- *
- * \param timer, the timer to start
- *
- */
-static void picoxcell_timer_start (int timer);
 
 /* Functions --------------------------------------------------------------- */
 
@@ -108,48 +98,6 @@ int checkboard (void)
 
 /*****************************************************************************
  *
- * print_cpuinfo()
- *
- * Purpose: Display some useful info about the cpu we are running on.
- *
- * Returns: 0 - Success
- *
- *****************************************************************************/
-int print_cpuinfo (void)
-{
-	unsigned int device_id, device_rev;
-
-	/* What device are we running on ? */
-	printf ("CPU:   ");
-
-	device_id = picoxcell_get_device_id ();	/* Read the device id */
-	device_rev = picoxcell_get_revision ();	/* Read the revision code */
-
-	switch (device_id) {
-	case PC302_DEVICE_ID:
-		printf ("PC302 Rev %04d\n", device_rev);
-		break;
-	case PC312_DEVICE_ID:
-		printf ("PC312 Rev %04d\n", device_rev);
-		break;
-	case PC313_DEVICE_ID:
-		printf ("PC313 Rev %04d\n", device_rev);
-		break;
-	case PC323_DEVICE_ID:
-		printf ("PC323 Rev %04d\n", device_rev);
-		break;
-	case PC333_DEVICE_ID:
-		printf ("PC333 Rev %04d\n", device_rev);
-		break;
-	default:
-		printf ("Unknown !\n");
-	}
-
-	return 0;
-}
-
-/*****************************************************************************
- *
  * misc_init_r()
  *
  * Purpose: Miscellaneous platform dependent initialisations
@@ -167,7 +115,7 @@ int misc_init_r (void)
  *
  * dram_init()
  *
- * Purpose: Initialize the DDR SDRAM info in the board data structure
+ * Purpose: Initialize the DDR SDRAM size info.
  *
  * Returns: 0 - Success
  *
@@ -179,51 +127,18 @@ int dram_init (void)
 	return 0;
 }
 
+/*****************************************************************************
+ *
+ * dram_init_banksize()
+ *
+ * Purpose: Initialize the the bd_t dram bank info.
+ *
+ * Returns: 0 - Success
+ *
+ *****************************************************************************/
 void dram_init_banksize (void)
 {
 	gd->bd->bi_dram[0].start = PHYS_SDRAM_1;
 	gd->bd->bi_dram[0].size =
 	    get_ram_size ((long *)PHYS_SDRAM_1, PHYS_SDRAM_1_SIZE);
-}
-
-static void picoxcell_timer_start (int timer)
-{
-	/* Make sure the timer is disabled */
-	picoxcell_write_register (0, (CONFIG_SYS_TIMERBASE +
-				      TIMERNCONTROLREGOFFSET (timer)));
-
-	/* Initialise the timer to all 1's.
-	 * We do this  because we want to run
-	 * the timer in free running mode.
-	 */
-	picoxcell_write_register (0xFFFFFFFF, (CONFIG_SYS_TIMERBASE +
-					       TIMERNLOADCOUNTREGOFFSET
-					       (timer)));
-
-	/* Start the timer in free running mode */
-	picoxcell_write_register ((TIMERINTERRUPTMASK | TIMERENABLE),
-				  (CONFIG_SYS_TIMERBASE +
-				   TIMERNCONTROLREGOFFSET (timer)));
-
-}
-
-/*****************************************************************************
- *
- * board_eth_init()
- *
- * Purpose: Initialize the Ethernet controller
- *
- * Returns: 0 - Success
- *
- *****************************************************************************/
-
-int board_eth_init (bd_t * bis)
-{
-	int ret = -1;
-
-#ifdef CONFIG_DW_EMAC
-	ret = picoxcell_eth_register (bis);
-#endif
-
-	return ret;
 }
